// src/os/src/types/dev/pci/class.ts
export enum KDriverPCIClass {
  MassStorageController = 1,
  NetworkController = 2,
  DisplayController = 3,
  MultimediaController = 4,
  MemoryController = 5,
  BridgeDevice = 6,
  SimpleCommunicationsController = 7,
  BaseSystemPeripheral = 8,
  InputDevice = 9,
  DockingStation = 10,
  Processors = 11,
  SerialBusController = 12,
  WirelessController = 13,
  IntelligentIOController = 14,
  SatelliteCommunicationController = 15,
  EncryptionDecryptionController = 16,
  DataAcquisitionAndSignalProcessingController = 17,
}